refactor(radio): refactor internal themed component
